Move task from modules to services
